@@ -1,4 +1,4 @@
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { FunctionComponent } from "react";
 
 import { Button, ButtonProps, ButtonType, ButtonVariant } from "../Button";
